move tests and benches to test-curves
