+++ ff/Cargo.toml
@@ -36,9 +36,7 @@ ark-test-curves = { workspace = true, features = [
     "secp256k1",
 ] }
 ark-algebra-test-templates = { path = "../test-templates" }
-p3_field = { git = "https://github.com/succinctlabs/Plonky3", package = "p3-field" }
-p3_goldilocks = { git = "https://github.com/succinctlabs/Plonky3", package = "p3-goldilocks" }
-blake2.workspace = true
+
 sha3.workspace = true
 sha2.workspace = true
 libtest-mimic.workspace = true
@@ -47,14 +45,6 @@ serde_json.workspace = true
 serde_derive.workspace = true
 hex.workspace = true
 criterion.workspace = true
-
-[[bench]]
-name = "cast_bench"
-harness = false
-
-[[bench]]
-name = "field_comparison"
-harness = false
 
 [features]
 default = []
