+++ test-curves/benches/smallfp_comparison.rs
@@ -1,40 +1,10 @@
-use ark_ff::ark_ff_macros::SmallFpConfig;
 use ark_ff::fields::{Fp128, Fp64, MontBackend, MontConfig};
-use ark_ff::{BigInt, SmallFp, SmallFpConfig, SqrtPrecomputation};
 use ark_std::hint::black_box;
+use ark_test_curves::smallfp128::SmallF128;
+use ark_test_curves::smallfp64::{SmallF64, SmallF64Mont};
 use criterion::{criterion_group, criterion_main, Criterion};
 use p3_field::AbstractField;
 use p3_goldilocks::Goldilocks as P3Goldilocks;
-
-// Goldilock's prime from Plonk3 18446744069414584321;
-
-#[derive(SmallFpConfig)]
-#[modulus = "18446744069414584321"]
-#[generator = "2"]
-#[backend = "standard"]
-pub struct SmallF64Config;
-pub type SmallF64 = SmallFp<SmallF64Config>;
-
-#[derive(SmallFpConfig)]
-#[modulus = "18446744069414584321"]
-#[generator = "2"]
-#[backend = "montgomery"]
-pub struct SmallF64ConfigMont;
-pub type SmallF64Mont = SmallFp<SmallF64ConfigMont>;
-
-#[derive(SmallFpConfig)]
-#[modulus = "18446744069414584321"]
-#[generator = "2"]
-#[backend = "standard"]
-pub struct SmallF128Config;
-pub type SmallF128 = SmallFp<SmallF128Config>;
-
-#[derive(SmallFpConfig)]
-#[modulus = "143244528689204659050391023439224324689"]
-#[generator = "2"]
-#[backend = "montgomery"]
-pub struct SmallF128ConfigMont;
-pub type SmallF128Mont = SmallFp<SmallF128ConfigMont>;
 
 #[derive(MontConfig)]
 #[modulus = "18446744069414584321"]
@@ -48,10 +18,7 @@ pub type F64 = Fp64<MontBackend<F64Config, 1>>;
 pub struct F128Config;
 pub type F128 = Fp128<MontBackend<F128Config, 2>>;
 
-// Benchmark functions
-
 fn naive_element_wise_mult_ark_bigint_64(c: &mut Criterion) {
-    // let's get four vectors and multiply them element-wise
     let len = 2_i64.pow(22);
     let v1: Vec<F64> = (0..len).map(F64::from).collect();
     let v2: Vec<F64> = (0..len).map(F64::from).collect();
@@ -73,7 +40,6 @@ fn naive_element_wise_mult_ark_bigint_64(c: &mut Criterion) {
 }
 
 fn naive_element_wise_mult_ark_small_field_64_std(c: &mut Criterion) {
-    // let's get four vectors and multiply them element-wise
     let len = 2_i64.pow(22);
     let v1: Vec<SmallF64> = (0..len).map(|i| SmallF64::from(i as u32)).collect();
     let v2: Vec<SmallF64> = (0..len).map(|i| SmallF64::from(i as u32)).collect();
@@ -95,7 +61,6 @@ fn naive_element_wise_mult_ark_small_field_64_std(c: &mut Criterion) {
 }
 
 fn naive_element_wise_mult_ark_small_field_64_mont(c: &mut Criterion) {
-    // let's get four vectors and multiply them element-wise
     let len = 2_i64.pow(22);
     let v1: Vec<SmallF64Mont> = (0..len).map(|i| SmallF64Mont::from(i as u32)).collect();
     let v2: Vec<SmallF64Mont> = (0..len).map(|i| SmallF64Mont::from(i as u32)).collect();
@@ -117,7 +82,6 @@ fn naive_element_wise_mult_ark_small_field_64_mont(c: &mut Criterion) {
 }
 
 fn naive_element_wise_mult_p3_64(c: &mut Criterion) {
-    // let's get four vectors and multiply them element-wise
     let len = 2_i64.pow(22);
     let v1: Vec<P3Goldilocks> = (0..len)
         .map(|i| P3Goldilocks::from_canonical_u64(i as u64))
@@ -148,7 +112,6 @@ fn naive_element_wise_mult_p3_64(c: &mut Criterion) {
 }
 
 fn naive_element_wise_mult_ark_bigint_128(c: &mut Criterion) {
-    // let's get four vectors and multiply them element-wise
     let len = 2_i64.pow(22);
     let v1: Vec<F128> = (0..len).map(F128::from).collect();
     let v2: Vec<F128> = (0..len).map(F128::from).collect();
