+++ test-curves/src/smallfp128.rs
@@ -1,0 +1,26 @@
+use ark_ff::ark_ff_macros::SmallFpConfig;
+use ark_ff::{BigInt, SmallFp, SmallFpConfig, SqrtPrecomputation};
+
+#[derive(SmallFpConfig)]
+#[modulus = "143244528689204659050391023439224324689"]
+#[generator = "3"]
+#[backend = "standard"]
+pub struct SmallF128Config;
+pub type SmallF128 = SmallFp<SmallF128Config>;
+
+#[derive(SmallFpConfig)]
+#[modulus = "143244528689204659050391023439224324689"]
+#[generator = "3"]
+#[backend = "montgomery"]
+pub struct SmallF128ConfigMont;
+pub type SmallF128Mont = SmallFp<SmallF128ConfigMont>;
+
+#[cfg(test)]
+mod tests {
+    use super::*;
+    use ark_algebra_test_templates::*;
+    use ark_std::vec;
+
+    test_small_field!(f128; SmallF128);
+    test_small_field!(f128_mont; SmallF128Mont);
+}
