+++ test-curves/src/smallfp64.rs
@@ -1,0 +1,26 @@
+use ark_ff::ark_ff_macros::SmallFpConfig;
+use ark_ff::{BigInt, SmallFp, SmallFpConfig, SqrtPrecomputation};
+
+#[derive(SmallFpConfig)]
+#[modulus = "18446744069414584321"] // Goldilock's prime 2^64 - 2^32 + 1
+#[generator = "7"]
+#[backend = "standard"]
+pub struct SmallF64Config;
+pub type SmallF64 = SmallFp<SmallF64Config>;
+
+#[derive(SmallFpConfig)]
+#[modulus = "18446744069414584321"] // Goldilock's prime 2^64 - 2^32 + 1
+#[generator = "7"]
+#[backend = "montgomery"]
+pub struct SmallF64ConfigMont;
+pub type SmallF64Mont = SmallFp<SmallF64ConfigMont>;
+
+#[cfg(test)]
+mod tests {
+    use super::*;
+    use ark_algebra_test_templates::*;
+    use ark_std::vec;
+
+    test_small_field!(f64; SmallF64);
+    test_small_field!(f64_mont; SmallF64Mont);
+}
